website/requirements/dev.txt
apipkg==1.5
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
black==19.10b0
boto==2.49.0
certifi==2019.11.28
cfgv==3.1.0
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
Django==3.0.7
django-haystack==2.8.1
django-markdownx==3.0.1
django-profanity-filter==0.2.1
django-simple-pagination==1.3
django-storages==1.9.1
django-webpack-loader==0.7.0
docopt==0.6.2
entrypoints==0.3
execnet==1.7.1
filelock==3.0.12
flake8==3.7.9
flake8-docstrings==1.5.0
gunicorn==20.0.4
identify==1.4.11
idna==2.9
importlib-metadata==1.5.0
importlib-resources==1.0.2
inflection==0.4.0
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
microurl==0.1.1
more-itertools==8.2.0
nodeenv==1.3.5
packaging==20.1
pathspec==0.7.0
Pillow==7.0.0
pluggy==0.13.1
pre-commit==2.1.1
psycopg2==2.8.5
psycopg2-binary==2.8.4
py==1.8.1
pycodestyle==2.5.0
pydocstyle==5.0.2
pyflakes==2.1.1
pylint==2.4.4
pyparsing==2.4.6
pytz==2019.3
PyYAML==5.3
regex==2020.2.20
requests==2.23.0
six==1.14.0
snowballstemmer==2.0.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==16.7.9
wcwidth==0.1.8
whitenoise==5.1.0
wrapt==1.11.2
zipp==2.2.0
